Add src/mock
